refactor: remove barrel files from app
