--- src/app/features/general/index.ts
@@ -1,1 +1,0 @@
-export * from './router'
