--- src/app/features/pokemon/application/dtos/index.ts
@@ -1,2 +1,0 @@
-export * from './get-pokemon-details-dto'
-export * from './get-pokemon-list-dto'
