--- src/app/features/pokemon/domain/models/index.ts
@@ -1,2 +1,0 @@
-export * from './pokemon-list-models'
-export * from './pokemon-details-models'
