--- src/app/features/pokemon/index.ts
@@ -1,1 +1,0 @@
-export * from './router'
