--- src/app/features/pokemon/presentation/components/index.ts
@@ -1,9 +1,0 @@
-export { PokeCard } from './PokeCard'
-export { PokeCardStat } from './PokeCardStat'
-export { PokeDetails } from './PokeDetails'
-export { PokeGrid } from './PokeGrid'
-export { PokeTabs } from './PokeTabs'
-export { PokemonNotFound } from './PokemonNotFound'
-export { SkeletonPokeCard } from './SkeletonPokeCard'
-export { SkeletonPokeList } from './SkeletonPokeList'
-export { PokeTypes } from './PokeTypes'
