--- src/app/features/pokemon/queries/index.ts
@@ -1,2 +1,0 @@
-export { pokemonDetailsQueryOptions } from './pokemon-details-query'
-export { pokemonListQueryOptions } from './pokemon-list-query'
